requirements.txt
streamlit
numpy
pandas
opencv-python
Pillow
scikit-learn
scikit-image
mysql-connector-python
